fix: restore architecture boundaries for nutrition capabilities
